- Inicio de creación de chats - JS de crear y borrar chats

+++ src/main/java/es/geeko/dto/ChatDto.java
@@ -16,8 +16,8 @@ import java.util.List;
 public class ChatDto {
 
     private long id;
-    private int idDestinatario;
-    private int idRemitente;
+    private long idDestinatario;
+    private long idRemitente;
     private String titulo;
     private String imagen;
     private String descripcion;
